Move `__main__.run_cli` to `cli.run_cli`
